refactor: move url to methods
